+++ crates/fuel-core/src/service/adapters.rs
@@ -1,16 +1,5 @@
-use crate::{
-    database::{
-        database_description::relayer::Relayer,
-        Database,
-    },
-    service::{
-        sub_services::{
-            BlockProducerService,
-            TxPoolSharedState,
-        },
-        vm_pool::MemoryPool,
-    },
-};
+use std::sync::Arc;
+
 use fuel_core_consensus_module::{
     block_verifier::Verifier,
     RelayerConsensusConfig,
@@ -24,13 +13,27 @@ use fuel_core_types::{
     services::block_importer::SharedImportResult,
 };
 use fuel_core_upgradable_executor::executor::Executor;
-use std::sync::Arc;
+
+use crate::{
+    database::{
+        database_description::relayer::Relayer,
+        Database,
+    },
+    service::{
+        sub_services::{
+            BlockProducerService,
+            TxPoolSharedState,
+        },
+        vm_pool::MemoryPool,
+    },
+};
 
 pub mod block_importer;
 pub mod consensus_module;
 pub mod consensus_parameters_provider;
 pub mod executor;
 pub mod fuel_gas_price_provider;
+pub mod gas_price_adapters;
 pub mod graphql_api;
 pub mod import_result_provider;
 #[cfg(feature = "p2p")]